Fix remaining CI lint and operator test failures
- Add manifest generation step to operator tests workflow
- Fix errcheck errors by explicitly ignoring viper.BindPFlag returns
- Remove unnecessary uint32 conversion in vrf.go
- Exclude lab directory from golangci-lint checks

+++ pkg/common/vrf/vrf.go
@@ -31,7 +31,7 @@ func Add(vpc, vpcAttachment string) error {
 		LinkAttrs: netlink.LinkAttrs{
 			Name: name,
 		},
-		Table: uint32(vrfId),
+		Table: vrfId,
 	}
 
 	if err := netlink.LinkAdd(vrf); err != nil {
